+++ app/src/main/AndroidManifest.xml
@@ -27,7 +27,7 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
-            android:name=".menu.MenuFolderActivity"
+            android:name=".menu.menuFolder.MenuFolderDetailActivity"
             android:exported="false" />
         <activity
             android:name=".menuinfo.MenuInfoActivity"
