+++ app/src/main/java/com/example/ourmenu/menu/menuFolder/MenuFolderDetailFragment.kt
@@ -1,31 +1,29 @@
-package com.example.ourmenu.menu
+package com.example.ourmenu.menu.menuFolder
 
 import android.graphics.RenderEffect
 import android.graphics.Shader
 import android.os.Build
 import android.os.Bundle
-import android.util.Log
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.annotation.RequiresApi
+import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.ourmenu.R
 import com.example.ourmenu.data.HomeMenuData
 import com.example.ourmenu.databinding.FragmentMenuFolderBinding
 import com.example.ourmenu.menu.adapter.MenuFolderRVAdapter
 
-class MenuFolderFragment : Fragment() {
-
+class MenuFolderDetailFragment : Fragment() {
     lateinit var binding: FragmentMenuFolderBinding
 
     @RequiresApi(Build.VERSION_CODES.S) // 이거 있어야 setRenderEffect 가능
     override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?,
     ): View? {
-
         binding = FragmentMenuFolderBinding.inflate(layoutInflater)
 
         initItemClickListener()
@@ -37,9 +35,7 @@ class MenuFolderFragment : Fragment() {
             editClicked()
         }
 
-
         return binding.root
-
     }
 
     private fun initItemClickListener() {
@@ -53,7 +49,7 @@ class MenuFolderFragment : Fragment() {
         val dummyItems = ArrayList<HomeMenuData>()
         for (i in 1..9) {
             dummyItems.add(
-                HomeMenuData(i.toString(), i.toString(), "store3")
+                HomeMenuData(i.toString(), i.toString(), "store3"),
             )
         }
 
@@ -63,7 +59,6 @@ class MenuFolderFragment : Fragment() {
 
     @RequiresApi(Build.VERSION_CODES.S)
     private fun editClicked() {
-
         // blur 효과 제거
         binding.clMenuFolderBlur.setRenderEffect(null)
         // 메뉴판 수정하기, 삭제하기, 취소 gone
@@ -71,7 +66,7 @@ class MenuFolderFragment : Fragment() {
 
         // 상단 이미지 blur 효과 적용
         binding.ivMenuFolderMainImage.setRenderEffect(
-            RenderEffect.createBlurEffect(2f, 2f, Shader.TileMode.CLAMP)
+            RenderEffect.createBlurEffect(2f, 2f, Shader.TileMode.CLAMP),
         )
         // vert 버튼 gone
         binding.ivMenuFolderVert.visibility = View.GONE
@@ -94,30 +89,29 @@ class MenuFolderFragment : Fragment() {
 
     @RequiresApi(Build.VERSION_CODES.S) // 이거 있어야 setRenderEffect 가능
     private fun initVertOnClickListener() {
-
         /* vert 버튼 클릭
-        * 배경에 blur 적용, 하단 뷰 visible */
+         * 배경에 blur 적용, 하단 뷰 visible */
         binding.ivMenuFolderVert.setOnClickListener {
             // 배경에 blur 효과 적용
             binding.clMenuFolderBlur.setRenderEffect(
-                RenderEffect.createBlurEffect(7.5f, 7.5f, Shader.TileMode.CLAMP)
+                RenderEffect.createBlurEffect(7.5f, 7.5f, Shader.TileMode.CLAMP),
             )
             // 메뉴판 수정하기, 삭제하기, 취소 visible
             binding.clMenuFolderVert.visibility = View.VISIBLE
         }
 
         /* 메뉴판 수정하기 클릭
-        * blur 효과 제거, 하단 뷰 gone
-        * 상단 이미지 blur 적용, vert 버튼 gone
-        * 카메라 layout visible
-        * edittext 설정 */
+         * blur 효과 제거, 하단 뷰 gone
+         * 상단 이미지 blur 적용, vert 버튼 gone
+         * 카메라 layout visible
+         * edittext 설정 */
         binding.btnMenuFolderEdit.setOnClickListener {
             editClicked()
         }
 
         /* 삭제하기 클릭
-        * blur 효과 제거, 하단 뷰 gone
-        * TODO 삭제하기 API */
+         * blur 효과 제거, 하단 뷰 gone
+         * TODO 삭제하기 API */
         binding.btnMenuFolderDelete.setOnClickListener {
             // blur 효과 제거
             binding.clMenuFolderBlur.setRenderEffect(null)
@@ -128,7 +122,7 @@ class MenuFolderFragment : Fragment() {
         }
 
         /* 취소 버튼 클릭
-        * blur 효과 제거, 하단 뷰 gone */
+         * blur 효과 제거, 하단 뷰 gone */
         binding.btnMenuFolderCancel.setOnClickListener {
             // blur 효과 제거
             binding.clMenuFolderBlur.setRenderEffect(null)
@@ -137,9 +131,9 @@ class MenuFolderFragment : Fragment() {
         }
 
         /* 확인 클릭
-        * 상단 이미지 blur 원상 복구, vert 버튼 visible
-        * edittext 원상 복구
-        * TODO 확인 버튼 클릭 API 구현 */
+         * 상단 이미지 blur 원상 복구, vert 버튼 visible
+         * edittext 원상 복구
+         * TODO 확인 버튼 클릭 API 구현 */
         binding.btnMenuFolderEditOk.setOnClickListener {
             // 상단 이미지 blur 효과 적용
             binding.ivMenuFolderMainImage.setRenderEffect(null)
@@ -157,9 +151,6 @@ class MenuFolderFragment : Fragment() {
             // 메뉴 추가하기 버튼 visible
             binding.btnMenuFolderAddMenu.visibility = View.VISIBLE
             // TODO 확인 버튼 클릭 API 구현
-
         }
-
     }
-
 }
