+++ app/src/main/res/layout/activity_menu_folder_detail.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".menu.MenuFolderActivity">
+    tools:context=".menu.menuFolder.MenuFolderDetailActivity">
 
     <FrameLayout
         android:id="@+id/menu_folder_frm"
