--- app/src/main/res/layout/fragment_menu.xml
@@ -1,148 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".menu.MenuFragment">
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:id="@+id/cl_menu_top_bar"
-        android:layout_width="match_parent"
-        android:layout_height="52dp"
-        android:background="@color/Neutral_White"
-        android:elevation="6dp"
-        app:layout_constraintTop_toTopOf="parent">
-
-        <ImageView
-            android:id="@+id/iv_menu_my_menu"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="20dp"
-            android:layout_marginTop="19dp"
-            android:src="@drawable/ic_my_menu"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
-        <ImageView
-            android:id="@+id/iv_menu_add_menu"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="18dp"
-            android:layout_marginEnd="20dp"
-            android:src="@drawable/ic_title_add"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-
-    <androidx.core.widget.NestedScrollView
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        app:layout_constraintTop_toBottomOf="@id/cl_menu_top_bar">
-
-        <androidx.constraintlayout.widget.ConstraintLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginBottom="20dp"
-            android:paddingBottom="20dp"
-            android:clipToPadding="false">
-
-            <Button
-                android:id="@+id/btn_menu_every_menu"
-                android:layout_width="match_parent"
-                android:layout_height="64dp"
-                android:layout_marginHorizontal="20dp"
-                android:layout_marginTop="76dp"
-                android:background="@drawable/button_shape_create"
-                app:backgroundTint="@null"
-                app:layout_constraintTop_toTopOf="parent" />
-
-            <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="10dp"
-                android:elevation="2dp"
-                android:fontFamily="@font/pretendard_700"
-                android:text="전체 메뉴 보기"
-                android:textColor="@color/Neutral_White"
-                android:textSize="18sp"
-                app:layout_constraintEnd_toEndOf="@id/btn_menu_every_menu"
-                app:layout_constraintStart_toStartOf="@id/btn_menu_every_menu"
-                app:layout_constraintTop_toTopOf="@id/btn_menu_every_menu" />
-
-            <LinearLayout
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="32dp"
-                android:elevation="2dp"
-                android:orientation="horizontal"
-                app:layout_constraintEnd_toEndOf="@id/btn_menu_every_menu"
-                app:layout_constraintStart_toStartOf="@id/btn_menu_every_menu"
-                app:layout_constraintTop_toTopOf="@id/btn_menu_every_menu">
-
-                <TextView
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:fontFamily="@font/pretendard_500"
-                    android:text="메뉴 "
-                    android:textColor="@color/Neutral_White"
-                    android:textSize="14sp" />
-
-                <TextView
-                    android:id="@+id/tv_menu_menu_number"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:fontFamily="@font/pretendard_500"
-                    android:text="0"
-                    android:textColor="@color/Neutral_White"
-                    android:textSize="14sp" />
-
-                <TextView
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:fontFamily="@font/pretendard_500"
-                    android:text="개"
-                    android:textColor="@color/Neutral_White"
-                    android:textSize="14sp" />
-
-
-            </LinearLayout>
-
-
-            <androidx.recyclerview.widget.RecyclerView
-                android:id="@+id/rv_menu_menu_folder"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:layout_marginTop="7dp"
-                android:layout_marginEnd="20dp"
-                android:orientation="vertical"
-                app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toBottomOf="@id/btn_menu_every_menu" />
-
-
-            <Button
-                android:id="@+id/btn_menu_add_menu_folder"
-                android:layout_width="match_parent"
-                android:layout_height="52dp"
-                android:layout_marginHorizontal="20dp"
-                android:background="@drawable/button_bg_add_menu_neutral_100"
-                android:elevation="6dp"
-                android:fontFamily="@font/pretendard_600"
-                android:text="메뉴판 추가하기"
-                android:textColor="@color/Neutral_500"
-                android:textSize="14sp"
-                android:paddingVertical="10dp"
-                app:backgroundTint="@null"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toBottomOf="@id/rv_menu_menu_folder" />
-
-            <!-- TODO 메뉴판 추가하기 버튼 누르면 UI 바뀌는것 7-18-->
-        </androidx.constraintlayout.widget.ConstraintLayout>
-
-    </androidx.core.widget.NestedScrollView>
-
-
-</androidx.constraintlayout.widget.ConstraintLayout>
